tienda/requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.8.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
Pygments==2.18.0
python-decouple==3.8
rich==13.9.3
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0